Adapted serial params to be a map from id to params Added static assert for incomplete type in Helper

+++ tests/Utilities/IPC/IPCTest.cpp
@@ -17,7 +17,7 @@
 struct Test
 {
 	int val1;
-	double val2;
+	float val2;
 	Vector3 val3;
 } test;
